+++ app/contact/page.js
@@ -4,7 +4,7 @@ import React, { useState } from "react";
 import { Button } from "@/components/ui/button";
 import { motion } from "framer-motion";
 import { Phone } from "lucide-react";
-import { FaWhatsapp, FaLinkedin } from "react-icons/fa";
+import { FaWhatsapp, FaLinkedin, FaInstagram } from "react-icons/fa";
 
 const Contact = () => {
   const [toastMessage, setToastMessage] = useState(null);
@@ -14,14 +14,55 @@ const Contact = () => {
     setTimeout(() => setToastMessage(null), duration);
   };
 
-  const handleSubmit = (e) => {
-    e.preventDefault();
+  const [formData, setFormData] = useState({
+    name: "",
+    email: "",
+    contact: "",
+    country: "",
+    message: "",
+  });
 
-    showToast(
-      "Message received",
-      "We'll get back to you as soon as possible.",
-      5000
-    );
+  const [loading, setLoading] = useState(false);
+  const [error, setError] = useState("");
+
+  const handleChange = (e) => {
+    setFormData({...formData, [e.target.name]: e.target.value});
+  };
+
+  const handleSubmit = async (e) => {
+    e.preventDefault();
+    setLoading(true);
+    setError("");
+
+    try {
+      const response = await fetch(
+        "https://medical-tourism-lqcu.onrender.com/api/patient/contact",
+        {
+          method: "POST",
+          headers: {
+            "Content-Type": "application/json",
+          },
+          body: JSON.stringify(formData),
+        }
+      );
+
+      const data = await response.json();
+
+      if(!response.ok) {
+        throw new Error(data.message || "Failed to Contact, Try Again!");
+      }
+
+      //After successful registration, show following message
+      showToast(
+        "Message received",
+        "We'll get back to you as soon as possible.",
+        5000
+      );
+    } catch (error) {
+      setError(error.message);
+    } finally {
+      setLoading(false);
+    }
 
     const form = e.target;
     form.reset();
@@ -160,15 +201,7 @@ const Contact = () => {
                 href="#"
                 className="w-10 h-10 rounded-full bg-white flex items-center justify-center text-primary hover:bg-primary hover:text-white transition-colors"
               >
-                <svg
-                  xmlns="http://www.w3.org/2000/svg"
-                  width="20"
-                  height="20"
-                  fill="currentColor"
-                  viewBox="0 0 16 16"
-                >
-                  <path d="M8 0C5.829 0 5.556.01 4.703.048 3.85.088 3.269.222 2.76.42a3.917 3.917 0 0 0-1.417.923A3.927 3.927 0 0 0 .42 2.76C.222 3.268.087 3.85.048 4.7.01 5.555 0 5.827 0 8 16s2.444-.01 3.298-.048c.851-.04 1.434-.174 1.943-.372a3.916 3.916 0 0 0 1.416-.923c.445-.445.718-.891.923-1.417.197-.509.332-1.09.372-1.942C15.99 10.445 16 10.173 16 8s-.01-2.445-.048-3.299c-.04-.851-.175-1.433-.372-1.941a3.926 3.926 0 0 0-.923-1.417A3.911 3.911 0 0 0 13.24.42c-.51-.198-1.092-.333-1.943-.372C10.443.01 10.172 0 7.998 0h.003zm-.717 1.442h.718c2.136 0 2.389.007 3.232.046.78.035 1.204.166 1.486.275.373.145.64.319.92.599.28.28.453.546.598.92.11.281.24.705.275 1.485.039.843.047 1.096.047 3.231s-.008 2.389-.047 3.232c-.035.78-.166 1.203-.275 1.485a2.47 2.47 0 0 1-.599.919c-.28.28-.546.453-.92.598-.28.11-.704.24-1.485.276-.843.038-1.096.047-3.232.047s-2.39-.009-3.233-.047c-.78-.036-1.203-.166-1.485-.276a2.478 2.478 0 0 1-.92-.598 2.48 2.48 0 0 1-.6-.92c-.109-.281-.24-.705-.275-1.485-.038-.843-.046-1.096-.046-3.233 0-2.136.008-2.388.046-3.231.036-.78.166-1.204.276-1.486.145-.373.319-.64.599-.92.28-.28.546-.453.92-.598.282-.11.705-.24 1.485-.276.738-.034 1.024-.044 2.515-.045v.002zm4.988 1.328a.96.96 0 1 0 0 1.92.96.96 0 0 0 0-1.92zm-4.27 1.122a4.109 4.109 0 1 0 0 8.217 4.109 4.109 0 0 0 0-8.217zm0 1.441a2.667 2.667 0 1 1 0 5.334 2.667 2.667 0 0 1 0-5.334z" />
-                </svg>
+                <FaInstagram />
               </a>
             </div>
           </motion.div>
@@ -180,6 +213,7 @@ const Contact = () => {
             transition={{ duration: 0.7 }}
             className="blue-glass rounded-2xl p-8 md:p-10 shadow-lg"
           >
+            {error && <p className="text-red-500"> {error} </p>}
             <form onSubmit={handleSubmit} className="space-y-6">
               <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
                 <div>
@@ -194,6 +228,7 @@ const Contact = () => {
                     id="name"
                     className="w-full px-4 py-3 rounded-lg border border-border focus:outline-none focus:ring-2 focus:ring-primary/20 focus:border-primary transition-all"
                     placeholder="John Doe"
+                    onChange={handleChange}
                     required
                   />
                 </div>
@@ -209,6 +244,7 @@ const Contact = () => {
                     id="email"
                     className="w-full px-4 py-3 rounded-lg border border-border focus:outline-none focus:ring-2 focus:ring-primary/20 focus:border-primary transition-all"
                     placeholder="john@example.com"
+                    onChange={handleChange}
                     required
                   />
                 </div>
@@ -227,6 +263,7 @@ const Contact = () => {
                     id="contact"
                     className="w-full px-4 py-3 rounded-lg border border-border focus:outline-none focus:ring-2 focus:ring-primary/20 focus:border-primary transition-all"
                     placeholder="Contact Number"
+                    onChange={handleChange}
                     required
                   />
                 </div>
@@ -242,6 +279,7 @@ const Contact = () => {
                     id="country"
                     className="w-full px-4 py-3 rounded-lg border border-border focus:outline-none focus:ring-2 focus:ring-primary/20 focus:border-primary transition-all"
                     placeholder="Your Country"
+                    onChange={handleChange}
                     required
                   />
                 </div>
@@ -259,15 +297,17 @@ const Contact = () => {
                   rows={4}
                   className="w-full px-4 py-3 rounded-lg border border-border focus:outline-none focus:ring-2 focus:ring-primary/20 focus:border-primary transition-all resize-none"
                   placeholder="Tell us about your medical needs..."
+                  onChange={handleChange}
                   required
                 ></textarea>
               </div>
 
               <Button
                 type="submit"
-                className="w-full rounded-lg py-6 text-base shadow-lg shadow-primary/20"
+                className="w-full rounded-lg py-6 bg-black text-base shadow-lg shadow-primary/20"
               >
-                Send Message
+                {loading ? "Sending Your Message": "Send Message"}
+                {/* Send Message */}
               </Button>
 
               <p className="text-xs text-muted-foreground text-center">
