+++ app/page.js
@@ -1,7 +1,7 @@
 // app/page.js
 import Hero from "@/components/Hero";
 import  WhyUs from "@/components/about";
-import Contact from "@/components/contact";
+import Contact from "./contact/page";
 import Dept from "@/components/departmentSec";
 import HowItWorks from "@/components/howItWork";
